[generate] refactor node scripts/generate to better support multiple-commands
